mal-dnssearch.sh: introduce new options and restructuring
Moved all log types and malware lists options into two options
where the type or list is specified as an argument. The result
of this is cleaner code and reduction of the number of options
which was excessive in my opinion options. This change intrduced
the ``-f'' option which specifies the log file to process, this
is required with ``-T <type>''.

The ability to specify multiple types and files at the same time
e.g. "...-p pdns.log -b dns.log" is no longer available. I don't
believe this is a critical feature anyway and you can just run
more than one instance of mal-dnssearch in a script to accomplish
the same thing.

``-f'' is now used to specify the filename, firewall blocking is
enabled by ``-F''.