mal-dnssearch.sh: check if file exits after download
Print message and exit when ``-N'' is used and the file isn't
in the current working directory. Sometimes a mal-list and forget
that I did, then I use ``-N'' and I get an ugly error that the file
doesn't exist.

From now on, warn the user properly, by checking that.